requirements.txt
pillow
piexif
pandas
requests
tqdm